services/github_scraper/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
PyGithub==2.1.1
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
typing-extensions==4.9.0
